chore: Restructure components
Bring back Header and Planet Patrol to Dashboard index.js
