--- src/components/Header/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import { Container, Wrapper, Title } from './styles';
-
-function Header () {
-  return (
-    <Container>
-      <Wrapper>
-        <Title>Planet Patrol</Title>
-  
-      </Wrapper>
-    </Container>
-  );
-}
-
-export default Header;
